Split further
