+++ Services/Quiz/Presenters/AnswerParser.cs
@@ -1,6 +1,6 @@
 using System.Text.RegularExpressions;
 
-namespace VocabularyTrainer.Services.Quiz;
+namespace VocabularyTrainer.Services.Quiz.Presenters;
 
 /// <summary>
 /// Pure static helpers for pre-processing answer strings before evaluation and hint display.
