+++ Services/Quiz/Presenters/SequenceAligner.cs
@@ -1,4 +1,4 @@
-namespace VocabularyTrainer.Services.Quiz;
+namespace VocabularyTrainer.Services.Quiz.Presenters;
 
 /// <summary>
 /// Pure sequence alignment: aligns <c>typed</c> against <c>correct</c> using edit distance,
