+++ Services/Quiz/QuizService.cs
@@ -1,6 +1,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using VocabularyTrainer.Models;
+using VocabularyTrainer.Services.Quiz.Distractors;
+using VocabularyTrainer.Services.Quiz.Presenters;
 using VocabularyTrainer.Services.Vocabulary;
 using QuizModel = VocabularyTrainer.Models.Quiz;
