+++ VocabularyTrainer.Tests/Services/ApplicationServiceTests.cs
@@ -4,6 +4,7 @@ using FluentAssertions;
 using VocabularyTrainer.Models;
 using VocabularyTrainer.Services;
 using VocabularyTrainer.Services.Quiz;
+using VocabularyTrainer.Services.Quiz.Distractors;
 using VocabularyTrainer.Services.Vocabulary;
 using VocabularyTrainer.Tests.Fakes;
 using VocabularyTrainer.Tests.Fixtures;
