+++ VocabularyTrainer.Tests/Services/Quiz/Distractors/EasyDistractorSelectorTests.cs
@@ -1,10 +1,10 @@
 using FluentAssertions;
 using VocabularyTrainer.Models;
-using VocabularyTrainer.Services.Quiz;
+using VocabularyTrainer.Services.Quiz.Distractors;
 using VocabularyTrainer.Tests.Fixtures;
 using Xunit;
 
-namespace VocabularyTrainer.Tests.Services.Quiz;
+namespace VocabularyTrainer.Tests.Services.Quiz.Distractors;
 
 public class EasyDistractorSelectorTests
 {
